Refactor LoginPortal
